feat: add supertest, cors and roll
