+++ jest.config.ts
@@ -3,5 +3,6 @@ export default {
   coverageDirectory: 'coverage',
   preset: 'ts-jest',
   testEnvironment: 'node',
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
   testPathIgnorePatterns: ['\\\\node_modules\\\\', 'dist/'],
 };
