+++ package.json
@@ -30,8 +30,11 @@
     "@commitlint/cli": "^12.1.4",
     "@commitlint/config-conventional": "^12.1.4",
     "@tsconfig/recommended": "^1.0.1",
+    "@types/cors": "^2.8.10",
     "@types/express": "^4.17.12",
     "@types/jest": "26.0.1",
+    "@types/roll": "^1.2.0",
+    "@types/supertest": "^2.0.11",
     "@typescript-eslint/eslint-plugin": "^4.28.1",
     "@typescript-eslint/parser": "^4.28.1",
     "eslint": "7.2.0",
@@ -43,12 +46,15 @@
     "prettier": "^2.3.2",
     "pretty-quick": "^3.1.1",
     "standard-version": "^9.3.0",
+    "supertest": "^6.1.3",
     "ts-jest": "^26.0.0",
     "ts-node": "^10.0.0",
     "ts-node-dev": "^1.1.8",
     "typescript": "^4.3.5"
   },
   "dependencies": {
-    "express": "^4.17.1"
+    "cors": "^2.8.5",
+    "express": "^4.17.1",
+    "roll": "^1.3.1"
   }
 }
