--- src/lib/sum.ts
@@ -1,4 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export default function sum(number1: number, number2: number): number {
-  return number1 + number2;
-}
