+++ tsconfig.json
@@ -6,5 +6,5 @@
     "moduleResolution": "node"
   },
   "typeAcquisition": { "include": ["jest"] },
-  "exclude": ["**/**.d.ts", "**/**.test.ts", "./jest.config.ts"]
+  "exclude": ["**/**.d.ts", "**/**.test.ts", "jest.config.ts"]
 }
